@@ -1,4 +1,5 @@
 ﻿
+using Microsoft.AspNetCore.Builder;
 using Microsoft.Extensions.DependencyInjection;
 using System;
 using Util.Caches;
@@ -10,7 +11,7 @@ namespace Microsoft.Extensions.DependencyInjection
     {
         /// <summary>
         /// 使用couchbase作为缓存
-        /// <para>注意：使用该方法时，还需要在服务注册完成之后，调用<see cref="CouchbaseExtensions.InitCouchBaseCacheClusterClient(IServiceProvider, Action{CouchbaseConfig})"/></para>
+        /// <para>注意：使用该方法时，还需要在服务注册完成之后，调用<see cref="CouchbaseExtensions.UseCouchBaseCache(IApplicationBuilder, Action{CouchbaseConfig})"/></para>
         /// </summary>
         /// <param name="services"></param>
         /// <param name="configurator">设置couchbase配置信息</param>
@@ -24,40 +25,40 @@ namespace Microsoft.Extensions.DependencyInjection
             return services;
         }
 
-        /// <summary>
-        /// 使用couchbase作为缓存，并对couchbase客户端进行初始化
-        /// </summary>
-        /// <param name="services"></param>
-        /// <param name="configurator">设置couchbase配置信息</param>
-        /// <returns></returns>
-        public static IServiceCollection AddCouchBaseCache(this IServiceCollection services, Action<CouchbaseConfig> configurator)
-        {
-            if (services == null)
-                throw new ArgumentNullException(nameof(services));
-            if (configurator == null)
-                throw new ArgumentNullException(nameof(configurator));
+        ///// <summary>
+        ///// 使用couchbase作为缓存，并对couchbase客户端进行初始化
+        ///// </summary>
+        ///// <param name="services"></param>
+        ///// <param name="configurator">设置couchbase配置信息</param>
+        ///// <returns></returns>
+        //public static IServiceCollection AddCouchBaseCache(this IServiceCollection services, Action<CouchbaseConfig> configurator)
+        //{
+        //    if (services == null)
+        //        throw new ArgumentNullException(nameof(services));
+        //    if (configurator == null)
+        //        throw new ArgumentNullException(nameof(configurator));
 
-            var config = new CouchbaseConfig();
-            configurator.Invoke(config);
-            if (config.Urls == null || config.Urls.Count == 0)
-                throw new Exception("couchbase 配置错误，必须提供url");
-            if (config.BucketAndPassword == null || config.BucketAndPassword.Count == 0)
-                throw new Exception("couchbase 配置错误，必须提供bucket");
+        //    var config = new CouchbaseConfig();
+        //    configurator.Invoke(config);
+        //    if (config.Urls == null || config.Urls.Count == 0)
+        //        throw new Exception("couchbase 配置错误，必须提供url");
+        //    if (config.BucketAndPassword == null || config.BucketAndPassword.Count == 0)
+        //        throw new Exception("couchbase 配置错误，必须提供bucket");
 
-            services.AddSingleton<ISectionCacheService>(new CouchbaseCacheService(config));
+        //    services.AddSingleton<ISectionCacheService>(new CouchbaseCacheService(config));
 
-            return services;
-        }
+        //    return services;
+        //}
 
         /// <summary>
         /// 初始化couchbase客户端
         /// </summary>
         /// <param name="serviceProvide"></param>
         /// <param name="configurator"></param>
-        public static void InitCouchBaseCacheClusterClient(IServiceProvider serviceProvide, Action<CouchbaseConfig> configurator)
+        public static void UseCouchBaseCache(this IApplicationBuilder app, Action<CouchbaseConfig> configurator)
         {
-            if (serviceProvide == null)
-                throw new ArgumentNullException(nameof(serviceProvide));
+            if (app == null)
+                throw new ArgumentNullException(nameof(app));
             if (configurator == null)
                 throw new ArgumentNullException(nameof(configurator));
 
@@ -68,7 +69,7 @@ namespace Microsoft.Extensions.DependencyInjection
             if (config.BucketAndPassword == null || config.BucketAndPassword.Count == 0)
                 throw new Exception("couchbase 配置错误，必须提供bucket");
 
-            var cacheService = serviceProvide.GetService<ICacheService>();
+            var cacheService = app.ApplicationServices.GetService<ICacheService>();
             if (cacheService is CouchbaseCacheService couchbaseCacheService)
                 couchbaseCacheService.InitClusterClient(config);
         }
