@@ -5,6 +5,7 @@ using Eagle.Application.Service;
 using Eagle.WebApi.Common;
 using Eagle.WebApi.EventHandlers.Events;
 using Eagle.WebApi.Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using Util.Caches;
@@ -40,12 +41,12 @@ namespace Eagle.WebApi.Controllers
             var log = Log.GetLog(this);
             log.Info("values get nlog");
             log.Info("values get nlog2");
-            return new string[] { "value1", "value2" };
+            return new string[] { "value1", "value2", "value3" };
         }
 
         // GET api/values/5
         [HttpGet("{id}")]
-        //[Authorize(Roles = "admin", Policy = "Permission")]
+        [Authorize(Roles = "admin", Policy = "Permission")]
         public async Task<ActionResult<string>> Get(int id)
         {
             if (id == 1)
@@ -74,6 +75,12 @@ namespace Eagle.WebApi.Controllers
             {
                 var systemConfig = _commonModuleService.GetSystemConfig();
                 return systemConfig.SystemKey;
+            }
+            if(id ==4)
+            {
+                await Task.Delay(30 * 1000);
+                //System.Threading.Thread.Sleep(30 * 1000);
+                return "delay value";
             }
             return "value";
         }
